Filter out pods/containers that are not running
The existing checks fail to detect Batch jobs that leave pods in a
completed state. This leads to an error message being logged as the
container does not exist and as such has no entry in the /sys/fs/cgroups
hierarchy. This change applies further checks to pods & containers to
identify ones that are running to avoid these error logs